feat: implement config display
