+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod config;
+pub mod nvy_config;
 pub mod export;
 pub mod init;
 pub mod log;
